support usb on cypherpunk version
