+++ src/tasks/ui_display_task.c
@@ -135,7 +135,6 @@ static void UiDisplayTask(void *argument)
 #endif
             }
             break;
-#ifdef WEB3_VERSION
             case UI_MSG_USB_TRANSPORT_VIEW: {
                 GuiFrameOpenViewWithParam(&g_USBTransportView, rcvMsg.buffer, rcvMsg.length);
             }
@@ -155,6 +154,7 @@ static void UiDisplayTask(void *argument)
                 }
             }
             break;
+#ifdef WEB3_VERSION
             case UI_MSG_CLOSE_NFT_LOCK: {
                 uint8_t *snapShotAddr = GetActSnapShot();
                 while (LcdBusy()) {
