+++ src/ui/gui_analyze/gui_analyze.c
@@ -952,7 +952,8 @@ void GuiAnalyzeViewInit(lv_obj_t *parent)
             yOffset = yOffset + lv_obj_get_content_height(child) + 16;
         }
         lv_obj_set_parent(g_analyzeTabview.obj[i], tabChild);
-    }
+        lv_obj_clear_flag(g_analyzeTabview.obj[i], LV_OBJ_FLAG_SCROLL_ELASTIC);
+   }
 }
 
 void *GuiTemplateReload(lv_obj_t *parent, uint8_t index)
