+++ src/ui/gui_views/gui_key_derivation_request_view.c
