+++ src/webusb_protocol/general/eapdu_protocol_parser.c
@@ -237,7 +237,6 @@ struct ProtocolParser *NewEApduProtocolParser()
 
 void GotoResultPage(EAPDUResultPage_t *resultPageParams)
 {
-#ifdef WEB3_VERSION
     if (resultPageParams != NULL) {
         if (GuiCheckIfTopView(&g_USBTransportView)) {
             return;
@@ -248,5 +247,4 @@ void GotoResultPage(EAPDUResultPage_t *resultPageParams)
             PubBufferMsg(UI_MSG_USB_TRANSPORT_VIEW, resultPageParams, sizeof(EAPDUResultPage_t));
         }
     }
-#endif
 }
